Rename adding _test in the programs name
