+++ test/package_ros/pkg_jetracer_move_test/src/brain_test.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-from jetracer import JetRacer
+from jetracer_test import JetRacer
 
 import rospy
 from geometry_msgs.msg import Twist
