+++ test/package_ros/pkg_rplidar_publisher_test/src/detect_obstacles_test.py
